fix(compras): compra em Guarani/Dólar volta a gerar pedido (integer vs. câmbio)
`invalid input syntax for type integer: "16.6"` ao transformar em pedido um
orçamento digitado em Gs. Em Reais funcionava.

Causa: todo valor de compras é gravado em centavos de BRL. Quando o orçamento
é digitado em outra moeda, a conversão pelo câmbio produz centavos quebrados —
Gs. 100 / 1204,82 = 8,3000 centavos. `purchase_budget_quotes.unit_price` já era
numeric(15,4) e guardava isso, mas ao gerar a OC o valor caía em colunas
INTEGER (purchase_order_items / supplier_order_items) e o Postgres recusava.

Arredondar não serve: 8,3 → 8 é 3,6% de erro no preço unitário, o que estoura
o total de uma compra inteira. Então:

- migration 062: unit_price/total_price de purchase_order_items e
  supplier_order_items, e unit_price de goods_receipt_items, passam a
  numeric(15,4) — mesma precisão do orçamento de origem. Os totais dos
  cabeçalhos seguem INTEGER (arredondamento < 1 centavo sobre a soma).
- mappers convertem com Number(): o driver pg devolve numeric como string e
  `subtotal += item.total_price` estava prestes a concatenar em vez de somar.
- generateSupplierOrderFromPO soma com Number() e arredonda antes de gravar.
- validação do PATCH de item do pedido aceita centavos fracionários.
- edição inline de preço no pedido do fornecedor passa a ser na moeda do
  orçamento, sem arredondar para centavo de Real.

Verificado ponta a ponta: orçamento PYG com cotação de 8,3 centavos
(Gs. 100/un × 60) gera OC e pedido com total 498 centavos = Gs. 6.000, e a
edição de item para 9,4567 grava 567,402 com subtotal 567.

+++ apps/api/src/controllers/supplier-orders.controller.ts
@@ -23,7 +23,9 @@ const GenerateOrdersSchema = z.object({
 
 const UpdateSupplierOrderItemSchema = z.object({
   quantity: z.number().int().positive().optional(),
-  unitPrice: z.number().int().min(0).optional(),
+  // Centavos de BRL — aceita fração porque compras em Gs/US$ convertem pelo
+  // câmbio e geram centavos quebrados (Gs. 100 = 8,3 centavos). Ver migration 062.
+  unitPrice: z.number().min(0).optional(),
   discountPercentage: z.number().min(0).max(100).optional(),
   notes: z.string().optional(),
 });
